Move code from lib.rs to line.rs
